requirements.txt
click==7.1.2
Flask==1.1.2
iniherit==0.3.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Werkzeug==1.0.1
yoyo-migrations==7.0.2
